Apply NEON-only S8 context storage and keep x86 quantized path unchanged
Agent-Logs-Url: https://github.com/bab2min/Kiwi/sessions/926e3fce-732d-4ef4-a9a3-0e31189bed2c

+++ src/CoNgramModel.hpp
@@ -111,12 +111,12 @@ namespace kiwi
 				return reinterpret_cast<const float*>(contextEmbPtr + idx * contextEmbStride());
 			}
 
-			inline const uint8_t* getContextQuantEmb(uint32_t idx) const
+			inline const uint8_t* getContextQuantEmb(size_t idx) const
 			{
 				return contextEmbPtr + idx * contextEmbStride();
 			}
 
-			inline const int8_t* getContextQuantEmbS8(uint32_t idx) const
+			inline const int8_t* getContextQuantEmbS8(size_t idx) const
 			{
 				return contextEmbS8Ptr + idx * (header.dim + sizeof(float));
 			}
